Updated Attribute to use DescendantsTracker
* Fixed bugs that this change exposed.
* Fixed Virtus.determine_type to lookup based on the primitive
* Clear out anonymous Attribute subclasses after specs
* Descendants are added to Attribute.descendants in a different order now,
  updated specs to use the correct order.
* Removed special cases from Virtus.determine_type specs

+++ spec/unit/virtus/attribute/numeric/class_methods/descendants_spec.rb
@@ -4,9 +4,9 @@ describe Virtus::Attribute::Numeric, '.descendants' do
   subject { described_class.descendants }
 
   let(:known_descendants) do
-    [ Virtus::Attribute::Decimal,
+    [ Virtus::Attribute::Integer,
       Virtus::Attribute::Float,
-      Virtus::Attribute::Integer ]
+      Virtus::Attribute::Decimal ]
   end
 
   it "should return all known attribute classes" do
